Redone for MVC - step 3
