+++ services/index.js
@@ -1,0 +1,3 @@
+const connectToContactsDB = require('./contactsMongoDB');
+
+module.exports = { connectToContactsDB };
